Hang on to your entities.
Every visible element will be classed as an entity and it is a subclass of an Entity
parent class. I am not quite finished with this, the game entities will all be stored
in a list and the gameworld class will be assigned to manage the list.

Also, I'll be adding to the way the game and window communicate with each other.
this should solve some problems I had with my 'Quick input handling hack'

-- Add : Entity types.
-- Add : UnukWindow
-- Add : Geometry - I have placed some vector stuff in here, to replace Vec2.*